Revert "Move `name_matches` to a separate file"
This reverts commit e9f20b70084f4137eb22bb722c34eb42c9ce3cde.
